Disable useless hostname verification checks on client certs
For some reason, starting with Jetty v. 9.4.15.v20190215, which is the
Jetty release included in Spring boot 2.1.4, the SSLContextFactory sets
configures the TLS socket in a way that results in an hostname check on
the client certificate. Hostname checks are good on the client side to
verify the TLS server certificates, it does not really make a lot of
sense for our use case of user client certificate (which are not bound
to any specific SAN or IP address).

So it is fine to disable the endpointIdentificationAlgorithm on our
server TLS sockets, as this was also the default in previous Jetty
versions.

The TLSServerConnectorBuilder has been enhanced to define what
endpointIdentificationAlgorithm should be used, with the default being
'no algorithm'.

This whole problem is irrelevant when the conscrypt provider is used
instead of the JSSE implementation.